fixed bug cause of new method "User[] getAll(Ctx,creds)" method

+++ com.openexchange.admin/test/com/openexchange/admin/rmi/UserTest.java
@@ -303,7 +303,7 @@ public class UserTest extends AbstractTest {
         
         boolean founduser = false;
         for(final User element: srv_response){
-            if(element.getId()==urs.getId()){
+            if(element.getId().intValue()==urs.getId().intValue()){
                 founduser = true;
             }
         }
